+++ CHANGELOG.md
@@ -11,6 +11,7 @@ Introduce LPSPI improvements:
 - Add additional checks for LPSPI frame sizes.
 - Rework LPSPI clock settings, initialization, and disable.
 - Add `flush()` method.
+- Add `soft_reset()` method.
 
 ## [0.5.4] 2023-11-26